[CK DSL] conv: address fourth review findings
Bug fixes:
- MlKernelConfig::enc_pipeline/dec_pipeline: add "basic_v1"->5 and
  "compv6"->6 entries; previously fell through to 0 ("compv3"),
  corrupting feature vectors for those pipeline variants
- kernels/gfx950/conv/manifest.json: add "pipeline":"mem"; kernel name
  contains _mem_cshuffle but no pipeline key caused enc_pipeline to
  default to "compv4" (enc=1) instead of "mem" (enc=3)
- ConvCandidateSweep: change xBytes/wBytes/yBytes from int32_t to
  int64_t; previous int32_t overflowed silently for large shapes
  (e.g. N=64, Hi=112, C=256 exceeds 2^31 fp16 bytes)

Narrowing fix:
- CkDslConvParamParser: add static_cast<int> for p.G assignment to
  silence int64_t -> int narrowing warning

Hardware profile:
- ConvHwProfile default (gfx950): num_cus 256->304, shader_engines
  32->38 to match MI350 silicon spec; prior value was copy-paste from
  the placeholder GEMM HardwareProfile

+++ dnn-providers/ck-dsl-provider/src/engines/CkDslConvParamParser.cpp
@@ -65,7 +65,7 @@ ParsedConvParams parseConvGraph(const hipdnn_flatbuffers_sdk::flatbuffer_utiliti
         const auto cpg = wd->Get(1);
         if (cpg <= 0 || p.C % cpg != 0)
             throw std::runtime_error("CkDslConv: weight dim[1] must be a positive divisor of C");
-        p.G = std::max((int64_t)1, p.C / cpg);
+        p.G = static_cast<int>(std::max((int64_t)1, p.C / cpg));
     }
     p.R  = wd->Get(2);
     p.S  = wd->Get(3);
